+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/item/NetworkItemHandler.java
@@ -10,24 +10,16 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.text.TextComponentTranslation;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class NetworkItemHandler implements INetworkItemHandler {
     private INetwork network;
 
-    private List<INetworkItem> items = new ArrayList<>();
-    private List<INetworkItem> itemsToRemove = new ArrayList<>();
+    private Map<EntityPlayer, INetworkItem> items = new ConcurrentHashMap<>();
 
     public NetworkItemHandler(INetwork network) {
         this.network = network;
-    }
-
-    @Override
-    public void update() {
-        items.removeAll(itemsToRemove);
-        itemsToRemove.clear();
     }
 
     @Override
@@ -57,31 +49,17 @@ public class NetworkItemHandler implements INetworkItemHandler {
         INetworkItem item = ((INetworkItemProvider) player.getHeldItem(hand).getItem()).provide(this, player, player.getHeldItem(hand));
 
         if (item.onOpen(network, player, hand)) {
-            items.add(item);
+            items.put(player, item);
         }
     }
 
     @Override
     public void onClose(EntityPlayer player) {
-        INetworkItem item = getItem(player);
-
-        if (item != null) {
-            itemsToRemove.add(item);
-        }
+        items.remove(player);
     }
 
     @Override
     public INetworkItem getItem(EntityPlayer player) {
-        Iterator<INetworkItem> it = items.iterator();
-
-        while (it.hasNext()) {
-            INetworkItem item = it.next();
-
-            if (item.getPlayer() == player) {
-                return item;
-            }
-        }
-
-        return null;
+        return items.get(player);
     }
 }